# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLAMP API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/mandates": {
      "get": {
        "summary": "List Mandates",
        "operationId": "list_mandates_api_mandates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Mandate"
                  },
                  "type": "array",
                  "title": "Response List Mandates Api Mandates Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Mandate",
        "operationId": "create_mandate_api_mandates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MandateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mandate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mandates/{mandate_id}": {
      "get": {
        "summary": "Get Mandate",
        "operationId": "get_mandate_api_mandates__mandate_id__get",
        "parameters": [
          {
            "name": "mandate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mandate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mandate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions": {
      "get": {
        "summary": "List Decisions",
        "operationId": "list_decisions_api_decisions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DecisionReceipt"
                  },
                  "type": "array",
                  "title": "Response List Decisions Api Decisions Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Decision",
        "operationId": "create_decision_api_decisions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions/{decision_id}": {
      "get": {
        "summary": "Get Decision",
        "operationId": "get_decision_api_decisions__decision_id__get",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions/{decision_id}/approve": {
      "post": {
        "summary": "Approve Decision",
        "operationId": "approve_decision_api_decisions__decision_id__approve_post",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions/{decision_id}/chain": {
      "post": {
        "summary": "Attach Chain",
        "operationId": "attach_chain_api_decisions__decision_id__chain_post",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChainAttachment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/metrics/summary": {
      "get": {
        "summary": "Metrics Summary",
        "operationId": "metrics_summary_api_metrics_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditPayload": {
        "properties": {
          "decision_id": {
            "type": "string",
            "title": "Decision Id"
          },
          "mandate_id": {
            "type": "string",
            "title": "Mandate Id"
          },
          "purchase_request_hash": {
            "type": "string",
            "title": "Purchase Request Hash"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionValue"
          },
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "decision_id",
          "mandate_id",
          "purchase_request_hash",
          "decision",
          "reason_code",
          "timestamp"
        ],
        "title": "AuditPayload"
      },
      "ChainAttachment": {
        "properties": {
          "network": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Network"
          },
          "tx_hash": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Tx Hash"
          }
        },
        "type": "object",
        "required": [
          "network",
          "tx_hash"
        ],
        "title": "ChainAttachment"
      },
      "DecisionCreate": {
        "properties": {
          "mandate_id": {
            "type": "string",
            "title": "Mandate Id"
          },
          "request": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Request"
          }
        },
        "type": "object",
        "required": [
          "mandate_id",
          "request"
        ],
        "title": "DecisionCreate"
      },
      "DecisionReceipt": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "decision_id": {
            "type": "string",
            "title": "Decision Id"
          },
          "mandate_id": {
            "type": "string",
            "title": "Mandate Id"
          },
          "original_request": {
            "type": "string",
            "title": "Original Request"
          },
          "structured_request": {
            "$ref": "#/components/schemas/StructuredPurchaseRequest"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionValue"
          },
          "matched_rule": {
            "type": "string",
            "title": "Matched Rule"
          },
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "blockchain_network": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockchain Network"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash"
          },
          "audit_payload": {
            "$ref": "#/components/schemas/AuditPayload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "decision_id",
          "mandate_id",
          "original_request",
          "structured_request",
          "decision",
          "matched_rule",
          "reason_code",
          "reason",
          "timestamp",
          "audit_payload"
        ],
        "title": "DecisionReceipt"
      },
      "DecisionValue": {
        "type": "string",
        "enum": [
          "ALLOW",
          "BLOCK",
          "NEEDS_HUMAN"
        ],
        "title": "DecisionValue"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Mandate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "total_budget": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Budget"
          },
          "remaining_budget": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Remaining Budget"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "allowed_merchants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Merchants"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "human_approval_threshold": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Human Approval Threshold"
          },
          "status": {
            "$ref": "#/components/schemas/MandateStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "purpose",
          "total_budget",
          "remaining_budget",
          "currency",
          "allowed_merchants",
          "expires_at",
          "human_approval_threshold",
          "status",
          "created_at"
        ],
        "title": "Mandate"
      },
      "MandateCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "purpose": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Purpose"
          },
          "total_budget": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Total Budget"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "allowed_merchants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Allowed Merchants"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "human_approval_threshold": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Human Approval Threshold"
          }
        },
        "type": "object",
        "required": [
          "name",
          "purpose",
          "total_budget",
          "currency",
          "allowed_merchants",
          "expires_at",
          "human_approval_threshold"
        ],
        "title": "MandateCreate"
      },
      "MandateStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "REVOKED",
          "EXPIRED"
        ],
        "title": "MandateStatus"
      },
      "StructuredPurchaseRequest": {
        "properties": {
          "merchant": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Merchant"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "item": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Item"
          },
          "purpose": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Purpose"
          }
        },
        "type": "object",
        "required": [
          "merchant",
          "amount",
          "currency",
          "item",
          "purpose"
        ],
        "title": "StructuredPurchaseRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}